criação da tela com botão e imagem
