+++ src/components/FeatureCarrocel/styles.scss
@@ -1,0 +1,76 @@
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;400;700&display=swap');
+
+$color-bg-blue: #D8E3F2;
+$color-pink: #C92071;
+$color-text: #474747;
+$color-rotulo: #E7FF86;
+$color-special: #F6AA1C;
+$font-inter: Inter, sans-serif;
+
+
+
+.Buttonpropscarro {
+    color: #D8E3F2;
+    background-color: $color-pink;
+    width: 200px;
+    height: 50px;
+    border-radius: 10px;
+    border: none;
+    font-weight: bold;
+    transition: 0.5s;
+    font-family: $font-inter;
+}
+
+.Buttonpropscarro:hover {
+    cursor: pointer;
+    background-color: #D8E3F2;
+    color: $color-pink;
+}
+
+
+.divcarousel {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+}
+
+
+.divinformation {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: flex-start;
+    
+    
+
+    h1 {        
+        font-style: normal;
+        font-weight: 800;
+        font-size: 64px;
+        line-height: 66px;
+        font-family: $font-inter;
+        letter-spacing: 1px;
+    }
+
+
+    p {
+        font-family: $font-inter;
+        font-style: normal;
+        font-weight: 400;
+        font-size: 18px;
+        line-height: 34px;
+
+    }
+
+    span {
+        color: $color-special;
+        font-weight: 700;
+        font-size: 16px;
+        line-height: 24px;
+        font-family: $font-inter;
+        letter-spacing: 0.75px;
+    }
+
+}
+
