+++ src/components/FeaturedCollections/styles.scss
@@ -58,8 +58,8 @@ div.propsText {
 
 .Buttonprops {
     color: $color-pink;
-    width: 130px;
-    height: 40px;
+    width: 150px;
+    height: 50px;
     border-radius: 10px;
     border: none;
     font-weight: bold;
